+++ config/settings/production.py
@@ -56,13 +56,6 @@ CSRF_COOKIE_SECURE = True
 CSRF_COOKIE_HTTPONLY = False  # False is required for AJAX to work
 X_FRAME_OPTIONS = "DENY"
 
-# SITE CONFIGURATION
-# ------------------------------------------------------------------------------
-# Hosts/domain names that are valid for this site
-# See https://docs.djangoproject.com/en/dev/ref/settings/#allowed-hosts
-ALLOWED_HOSTS = env.list("DJANGO_ALLOWED_HOSTS", default=["varfish.bihealth.org"])
-# END SITE CONFIGURATION
-
 INSTALLED_APPS += ["gunicorn"]
 
 
@@ -149,46 +142,6 @@ CACHES = {
     }
 }
 
-
-if env("DJANGO_SENTRY_DSN", default=None):
-    # Sentry Configuration
-    SENTRY_DSN = env("DJANGO_SENTRY_DSN")
-    SENTRY_CLIENT = env(
-        "DJANGO_SENTRY_CLIENT", default="raven.contrib.django.raven_compat.DjangoClient"
-    )
-    LOGGING = {
-        "version": 1,
-        "disable_existing_loggers": True,
-        "root": {"level": "WARNING", "handlers": ["sentry"]},
-        "formatters": {
-            "verbose": {
-                "format": "%(levelname)s %(asctime)s %(module)s "
-                "%(process)d %(thread)d %(message)s"
-            }
-        },
-        "handlers": {
-            "sentry": {
-                "level": "ERROR",
-                "class": "raven.contrib.django.raven_compat.handlers.SentryHandler",
-            },
-            "console": {"level": "DEBUG", "class": "logging.StreamHandler", "formatter": "verbose"},
-        },
-        "loggers": {
-            "django.db.backends": {"level": "ERROR", "handlers": ["console"], "propagate": False},
-            "raven": {"level": "DEBUG", "handlers": ["console"], "propagate": False},
-            "sentry.errors": {"level": "DEBUG", "handlers": ["console"], "propagate": False},
-            "django.security.DisallowedHost": {
-                "level": "ERROR",
-                "handlers": ["console", "sentry"],
-                "propagate": False,
-            },
-        },
-    }
-    SENTRY_CELERY_LOGLEVEL = env.int("DJANGO_SENTRY_LOG_LEVEL", logging.INFO)
-    RAVEN_CONFIG = {
-        "CELERY_LOGLEVEL": env.int("DJANGO_SENTRY_LOG_LEVEL", logging.INFO),
-        "DSN": SENTRY_DSN,
-    }
 
 # Custom Admin URL, use {% url 'admin:index' %}
 # ADMIN_URL = env("DJANGO_ADMIN_URL")