+++ HISTORY.rst
@@ -20,6 +20,7 @@ End-User Summary
 - Fixed bug that lead to an AJAX error in the filter view when previous filter results failed to load because the variants of a case were deleted in the meantime.
 - Entering the filter view is now only possible when there are variants and a variant set.
   When there are variant reported but no variant set, a warning in form of a small red icon next to the number of variants is displayed, complaining about an inconsistent state.
+- In case of errors, you can now give feedback in a form via Sentry.
 
 Full Change List
 ================
@@ -39,6 +40,7 @@ Full Change List
 - Fixed bug that lead to an AJAX error in the filter view when previous filter results failed to load because the variants of a case were deleted in the meantime.
 - Entering the filter view is now only possible when there are variants and a variant set.
   When there are variant reported but no variant set, a warning in form of a small red icon next to the number of variants is displayed, complaining about an inconsistent state.
+- Using latest sentry SDK client.
 
 -------
 v0.19.0